Extra Info
It works, need to add some more info as parts of the game get added, but for the current state of the game it displays the info that is probably most relevant. There is one slight bug where when the local multiplayer scene first opens the label is in the wrong spot, will try to fix that later.